+++ agencies/serializers.py
@@ -73,6 +73,7 @@ def serialize_agency(agency, user):
         "isHidden": agency.is_hidden,
         "mapColor": agency.map_color,
         "zeroDayPool": agency.zero_day_pool if is_admin else None,
+        "sweepPool": agency.sweep_pool,
         "conditions_cyber": [
             {
                 "id": c.id,